Relocate kmeans to integrations
